% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_isco88}
\alias{add_isco88}
\title{Add ISCO-88 occupation codes}
\usage{
add_isco88(ukb_data, soc2000_var)
}
\arguments{
\item{ukb_data}{A data frame containing UK Biobank data.}

\item{soc2000_var}{A character string indicating the column in \code{ukb_data} that contains the SOC2000 codes (i.e. \code{job_code_at_visit_f20277_0_0}).}
}
\value{
A data frame with the UK Biobank data merged with the ISCO-88 occupation, based on the SOC2000 codes.
}
\description{
Matches SOC2000 codes to ISCO-88, the 1988 version of the International Standard Classification of Occupations.
}
\details{
ISCO-88 is a standardized classification of occupations used internationally.
It provides a comprehensive and systematic way of classifying different jobs.
The ISCO-88 classification has four levels, ranging from the most general to the most specific and
its codes are widely used in social sciences, organizational psychology and human resource management.
By matching SOC2000 codes to ISCO-88, this function allows the researcher to link the UKB to
data with the internationally most common standardized occupation classification system.
The mapping is based on the .sps-file provided at http://www.camsis.stir.ac.uk/occunits/uksoc2000toisco88v3.sps.
}
