% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{add_traits}
\alias{add_traits}
\title{Add cognitive and non-cognitive traits}
\usage{
add_traits(ukb_data, soc2000_var)
}
\arguments{
\item{ukb_data}{A data frame containing UK Biobank data.}

\item{soc2000_var}{A character string indicating the column in \code{ukb_data} that contains the SOC2000 codes (i.e. \code{job_code_at_visit_f20277_0_0}).}
}
\value{
A data frame with the UK Biobank data merged with the information on cognition and non-cognitive traits.
}
\description{
Matches SOC2000 codes to british occupational means of cognition and non-cognitive traits.
}
\details{
Wolfram (2023) estimated occupational means of cognitive ability, the big 5 personality dimensions, risk taking, mental health,
gratification delay and self efficacy for 360 occupations in Great Britain using the SOC2010.
By relying on the official crosswalks, we converted these values back to SOC2000.

Wolfram, T. (2023). (Not just) Intelligence stratifies the occupational hierarchy: Ranking 360 professions by IQ and non-cognitive traits. Intelligence, 98, 101755.
}
